+++ src/constraints/battery_degradation.jl
@@ -11,7 +11,7 @@ function add_degradation_variables(m, p)
 end
 
 
-function constrain_degradation_variables(m, p; b="ElectricStorage")
+function constrain_degradation_variables(m, p, b)
     days = 1:365*p.s.financial.analysis_years
     ts_per_day = 24 / p.hours_per_time_step
     ts_per_year = ts_per_day * 365
@@ -44,14 +44,12 @@ end
 
 
 """
-    add_degradation(m, p; b="ElectricStorage")
+    add_degradation(m, p, b)
 
 NOTE the average SOC and EFC variables are in absolute units. For example, the SOH variable starts 
     at the battery capacity in kWh.
 """
-function add_degradation(m, p; b="ElectricStorage")
-    
-    # Indices
+function add_degradation(m, p, b)
     days = 1:365*p.s.financial.analysis_years
     months = 1:p.s.financial.analysis_years*12
 
@@ -75,7 +73,7 @@ function add_degradation(m, p; b="ElectricStorage")
     end
 
     add_degradation_variables(m, p)
-    constrain_degradation_variables(m, p, b=b)
+    constrain_degradation_variables(m, p, b)
 
     @constraint(m, [d in 2:days[end]],
         m[:SOH][d] == m[:SOH][d-1] - p.hours_per_time_step * (