+++ CHANGELOG.md
@@ -25,6 +25,25 @@ Classify the change according to the following categories:
     ### Deprecated
     ### Removed
 
+## Develop LDES
+### Added
+- Added input option **can_export_to_grid** (defaults to _false_) to `ElectricStorage` and decision variable **dvStorageToGrid**
+- Added input option **optimize_soc_init_fraction** (defaults to _false_) to `ElectricStorage`, which makes the optimization choose the inital SOC (equal to final SOC) instead of using **soc_init_fraction**. The initial SOC is also constrained to equal the final SOC, which eliminates the "free energy" issue. We currently do not fix SOC when **soc_init_fraction** is used because this has caused infeasibility. 
+- Added output **initial_capital_cost** to all techs
+- Added input **soc_based_per_ts_thermal_decay_fraction** to both `ColdThermalStorage` and `HotThermalStorage` for modeling SOC-based thermal decay rate 
+- Renamed **thermal_decay_rate_fraction** to **capacity_based_per_ts_thermal_decay_fraction** in both `ColdThermalStorage` and `HotThermalStorage` to clarify that the variable represents the capacity-based thermal decay rate
+- Added ability to model multiple `ElectricStorage` at once. Inputs are provided as a _list_ of _dict_ under the `ElectricStorage` key and outputs are similarly formatted. This mirrors the existing multiple `PV` implementation. This functionality is not yet implemented for MPC scenarios.
+### Fixed
+- Added missing outputs **lifecycle_export_benefit_before_tax** and **year_one_export_benefit_after_tax** to `ElectricTariff`
+- Add missing output **year_one_om_cost_before_tax** to `PV`
+
+## Develop battery_om
+### Added
+- Added inputs **om_cost_per_kw** and **om_cost_per_kwh** to `ElectricStorage` for modeling capacity-based O&M 
+- Added outputs **lifecycle_om_cost_after_tax** and **year_one_om_cost_before_tax** to `ElectricStorage` 
+- Added input **soc_based_per_ts_self_discharge_fraction** to `ElectricStorage` for modeling SOC-based battery self-discharge
+- Added input **capacity_based_per_ts_self_discharge_fraction** to `ElectricStorage` for modeling capacity-based battery self-discharge
+
 ## Develop - 2024-10-11
 ### Added
 - Add new **ElectricStorage** parameters **max_duration_hours** and **min_duration_hours** to bound the energy duration of battery storage